WebApp/requirements.txt
flask
flask-cors
python-dotenv
langchain-groq
langchain
elevenlabs